refactor(distributed): rename split package to dataunit and remove unused operator factories
- Rename split/ package to dataunit/ in both core and opensearch modules
- Delete SourceOperatorFactory, OperatorFactory, and Pipeline (unused)
- Simplify ComputeStage constructor by removing factory fields
- Update all imports across 10+ files
